+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+<<<<<<< HEAD
 ### v2.8.7 
 *2014-01-19*
+=======
+**MayhemYDG**:
+- Added a `Reset Settings` button in the settings.
+- More stability update.
+- Stability update.
+>>>>>>> v3
 
 **Zixaphir**:
 - Fix posting.